feat(api): SHR-177 extract diffs router to server/routes/diffs.ts
Move GET /api/tasks/:id/diff, GET /api/tasks/:id/diff/*path,
GET /api/tasks/:id/branches, GET /api/tasks/:id/commits,
PATCH /api/tasks/:id/base into dedicated diffs router (5 routes).
Remove now-unused imports from api.ts.
