+++ server/api.ts
@@ -17,39 +17,9 @@ import {
   createShellTerminal,
   closeShellTerminal,
 } from './task-runner.js';
-import * as diffMod from './diff.js';
-import { taskWorkingDir } from './task-paths.js';
-import { parseDiffRange } from './diff-range.js';
-import {
-  BaseBranchMissingError,
-  BaseUnavailableError,
-  clearDiffBaseCache,
-  resolveDiffBase,
-  resolveRef,
-} from './diff-base.js';
-import { listBranches, listCommits } from './git-commits.js';
-import { setReviewed, clearReviewed } from './repositories/file-review-state.js';
-import {
-  listComments,
-  getComment,
-  resolveComment,
-  unresolveComment,
-  updateCommentBody,
-  deleteComment,
-  seedInlineComment,
-} from './repositories/inline-comments.js';
-import { computeOutdated } from './inline-comments-outdated.js';
 import { sendMessageToAgent } from './tmux-input.js';
-import { listReviewsInbox, getReviewDetail } from './reviews-inbox.js';
-import {
-  getReviewRun,
-  getCurrentRun,
-  setWalkthrough,
-  seedReviewRun,
-} from './repositories/review-runs.js';
-import { listPublishedReviews } from './repositories/published-reviews.js';
-import { addLearning } from './repositories/review-learnings.js';
-import { updateCommentFields } from './repositories/inline-comments.js';
+import { seedInlineComment } from './repositories/inline-comments.js';
+import { seedReviewRun } from './repositories/review-runs.js';
 import { mountArtifactEndpoint } from './orchestrator/artifact-endpoint.js';
 
 import { validateAgentName } from './harnesses/types.js';
@@ -85,8 +55,6 @@ import {
   hardDeleteTask,
   touchLastViewed,
   touchAllLastViewed,
-  touchUpdatedAt,
-  setWorktreeBase,
   setCurrentSummary,
   addTaskUpdate,
   listTaskUpdates,
@@ -94,7 +62,6 @@ import {
   getTaskExternalRef,
   upsertTaskExternalRef,
   deleteTaskExternalRef,
-  findExistingReviewTask,
   getWorktree,
   listWorktrees as listWorktreesRepo,
   listTasksByWorktree,
@@ -111,19 +78,12 @@ import {
   listUserTerminals,
   listUserTerminalsByTasks,
   getUserTerminalByIdAndTask,
-  listTrackedRepoPaths,
   insertWorktreeIfAbsent,
   insertTaskIfAbsent,
   inTransaction,
 } from './repositories/index.js';
 
-import { createInlineComment } from './services/comment-service.js';
 import { createTask } from './services/task-service.js';
-import {
-  createReviewTaskFromPr,
-  createManualReview,
-  triggerReviewRun,
-} from './services/review-service.js';
 import { ServiceError } from './services/errors.js';
 
 import { router as miscRouter } from './routes/misc.js';
@@ -137,11 +97,14 @@ import { router as chatsRouter } from './routes/chats.js';
 import { router as agentDefsRouter } from './routes/agent-defs.js';
 import { router as orchestratorRouter } from './routes/orchestrator.js';
 import { router as integrationsRouter } from './routes/integrations.js';
+import { router as reviewsRouter } from './routes/reviews.js';
+import { router as reviewRunsRouter } from './routes/review-runs.js';
+import { router as commentsRouter } from './routes/comments.js';
+import { router as diffsRouter } from './routes/diffs.js';
 
 import {
   loadTaskOrFail,
   lookupExistingReviewId,
-  deepMerge,
   fetchTaskBundle,
   derivedStatus,
 } from './routes/_shared.js';
@@ -165,6 +128,10 @@ export function setupRoutes(app: Express): void {
   app.use(agentDefsRouter);
   app.use(orchestratorRouter);
   app.use(integrationsRouter);
+  app.use(reviewsRouter);
+  app.use(reviewRunsRouter);
+  app.use(commentsRouter);
+  app.use(diffsRouter);
 
   // List all tasks with agents
   app.get('/api/tasks', (req: Request, res: Response) => {
@@ -777,555 +744,6 @@ export function setupRoutes(app: Express): void {
     res.json({ success: true });
   });
 
-  // ─── Diff ──────────────────────────────────────────────────────────────────
-
-  app.get('/api/tasks/:id/diff', async (req: Request, res: Response) => {
-    const task = loadTaskOrFail(req, res);
-    if (!task) return;
-    if (task.run_mode === 'scratch') {
-      res.status(400).json({ error: 'no repo for scratch task' });
-      return;
-    }
-    const cwd = taskWorkingDir(task);
-    if (!cwd) {
-      res.status(400).json({ error: 'Task has no worktree' });
-      return;
-    }
-    if (!task.base_sha) {
-      res.status(400).json({ error: 'base_sha not available for this task' });
-      return;
-    }
-    if (!fs.existsSync(cwd)) {
-      res.status(400).json({ error: 'Worktree no longer exists on disk' });
-      return;
-    }
-    let range;
-    try {
-      range = parseDiffRange(typeof req.query.range === 'string' ? req.query.range : undefined);
-    } catch (err) {
-      res.status(400).json({ error: (err as Error).message });
-      return;
-    }
-    try {
-      const summary = await diffMod.getDiffSummary({ task, range });
-      res.json(summary);
-    } catch (err) {
-      if (err instanceof BaseBranchMissingError) {
-        res.status(422).json({ error: 'base_branch_missing', message: err.message });
-        return;
-      }
-      if (err instanceof BaseUnavailableError) {
-        res.status(503).json({ error: 'base_unavailable', message: err.message });
-        return;
-      }
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  app.get('/api/tasks/:id/diff/*path', async (req: Request, res: Response) => {
-    const task = loadTaskOrFail(req, res);
-    if (!task) return;
-    if (task.run_mode === 'scratch') {
-      res.status(400).json({ error: 'no repo for scratch task' });
-      return;
-    }
-    const cwd = taskWorkingDir(task);
-    if (!cwd) {
-      res.status(400).json({ error: 'Task has no worktree' });
-      return;
-    }
-    if (!task.base_sha) {
-      res.status(400).json({ error: 'base_sha not available for this task' });
-      return;
-    }
-    if (!fs.existsSync(cwd)) {
-      res.status(400).json({ error: 'Worktree no longer exists on disk' });
-      return;
-    }
-    const params = req.params as Record<string, string | string[]>;
-    const rawPath = params.path ?? params['0'] ?? '';
-    const relPath = Array.isArray(rawPath) ? rawPath.join('/') : rawPath;
-    try {
-      diffMod.safeResolvePath(cwd, relPath);
-    } catch {
-      res.status(400).json({ error: 'Invalid path' });
-      return;
-    }
-    let range;
-    try {
-      range = parseDiffRange(typeof req.query.range === 'string' ? req.query.range : undefined);
-    } catch (err) {
-      res.status(400).json({ error: (err as Error).message });
-      return;
-    }
-    try {
-      // Resolve the live base so the per-file diff agrees with the summary
-      // (both go through resolveDiffBase, which shares an in-process cache).
-      const resolved = await resolveDiffBase(task);
-      const diff = await diffMod.getFileDiff({
-        worktree: cwd,
-        range,
-        taskBaseSha: resolved.sha,
-        relPath,
-      });
-      res.json(diff);
-    } catch (err) {
-      if (err instanceof BaseBranchMissingError) {
-        res.status(422).json({ error: 'base_branch_missing', message: err.message });
-        return;
-      }
-      if (err instanceof BaseUnavailableError) {
-        res.status(503).json({ error: 'base_unavailable', message: err.message });
-        return;
-      }
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  // ─── Branches / commits / base mutation ─────────────────────────────────────
-
-  app.get('/api/tasks/:id/branches', async (req: Request, res: Response) => {
-    const task = loadTaskOrFail(req, res);
-    if (!task) return;
-    if (task.run_mode === 'scratch') {
-      res.status(400).json({ error: 'no repo for scratch task' });
-      return;
-    }
-    const cwd = taskWorkingDir(task);
-    if (!cwd || !fs.existsSync(cwd)) {
-      res.status(400).json({ error: 'Task has no usable worktree' });
-      return;
-    }
-    try {
-      const result = await listBranches(cwd);
-      res.json(result);
-    } catch (err) {
-      apiLogger.warn({ task_id: task.id, err: (err as Error).message }, 'listBranches failed');
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  app.get('/api/tasks/:id/commits', async (req: Request, res: Response) => {
-    const task = loadTaskOrFail(req, res);
-    if (!task) return;
-    if (task.run_mode === 'scratch') {
-      res.status(400).json({ error: 'no repo for scratch task' });
-      return;
-    }
-    const cwd = taskWorkingDir(task);
-    if (!cwd || !fs.existsSync(cwd)) {
-      res.status(400).json({ error: 'Task has no usable worktree' });
-      return;
-    }
-
-    // Determine the from/to refs. If `range=` is provided, derive from it; else
-    // default to base..HEAD when we have a base_sha, or just HEAD when we don't.
-    let from: string | undefined;
-    let to = 'HEAD';
-    const rangeParam = typeof req.query.range === 'string' ? req.query.range : undefined;
-    if (rangeParam) {
-      try {
-        const parsed = parseDiffRange(rangeParam);
-        switch (parsed.kind) {
-          case 'base':
-            from = task.base_sha ?? undefined;
-            to = 'HEAD';
-            break;
-          case 'commit':
-            from = `${parsed.sha}^`;
-            to = parsed.sha;
-            break;
-          case 'range':
-            from = parsed.from;
-            to = parsed.to;
-            break;
-          case 'working':
-            // No commits in a working-only range.
-            res.json({ commits: [], truncated: false });
-            return;
-        }
-      } catch (err) {
-        res.status(400).json({ error: (err as Error).message });
-        return;
-      }
-    } else if (task.base_sha) {
-      from = task.base_sha;
-    }
-
-    const limitRaw = Number(req.query.limit ?? 200);
-    const limit = Math.min(Math.max(Number.isFinite(limitRaw) ? limitRaw : 200, 1), 1000);
-
-    try {
-      const result = await listCommits(cwd, { from, to, limit });
-      res.json(result);
-    } catch (err) {
-      apiLogger.warn({ task_id: task.id, err: (err as Error).message }, 'listCommits failed');
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  app.patch('/api/tasks/:id/base', async (req: Request, res: Response) => {
-    const task = loadTaskOrFail(req, res);
-    if (!task) return;
-    if (task.run_mode === 'scratch') {
-      res.status(400).json({ error: 'no repo for scratch task' });
-      return;
-    }
-    if (task.runtime_state === 'idle') {
-      res.status(409).json({ error: 'cannot change base on a draft task' });
-      return;
-    }
-    if (!task.worktree_id) {
-      res.status(400).json({ error: 'task has no worktree row to update' });
-      return;
-    }
-    const cwd = taskWorkingDir(task);
-    if (!cwd || !fs.existsSync(cwd)) {
-      res.status(400).json({ error: 'Task has no usable worktree' });
-      return;
-    }
-
-    const baseBranch = (req.body as { base_branch?: unknown }).base_branch;
-    if (typeof baseBranch !== 'string' || !baseBranch.trim()) {
-      res.status(400).json({ error: 'base_branch is required' });
-      return;
-    }
-
-    let sha: string;
-    try {
-      sha = await resolveRef(cwd, baseBranch);
-    } catch (err) {
-      apiLogger.warn(
-        { task_id: task.id, base_branch: baseBranch, err: (err as Error).message },
-        'resolveRef failed',
-      );
-      res.status(400).json({ error: 'ref does not resolve' });
-      return;
-    }
-
-    // Persist the new base on the joined worktrees row (Phase 2a moved these
-    // columns off `tasks`). Bump the task's updated_at separately.
-    setWorktreeBase(task.worktree_id, baseBranch, sha);
-    touchUpdatedAt(task.id);
-
-    // Invalidate any cached origin tip for old/new branch on this worktree so
-    // the next diff fetch resolves fresh.
-    if (task.base_branch) clearDiffBaseCache(cwd, task.base_branch);
-    clearDiffBaseCache(cwd, baseBranch);
-
-    apiLogger.info(
-      { task_id: task.id, base_branch: baseBranch, base_sha: sha },
-      'task base changed',
-    );
-
-    broadcast({ type: 'task:updated', payload: { taskId: task.id } });
-
-    const reloaded = getTaskRepo(task.id);
-    res.json(reloaded);
-  });
-
-  // ─── File review state ─────────────────────────────────────────────────────
-
-  app.post('/api/tasks/:id/files/*path/reviewed', async (req: Request, res: Response) => {
-    const task = loadTaskOrFail(req, res);
-    if (!task) return;
-    const cwd = taskWorkingDir(task);
-    if (!cwd) {
-      res.status(400).json({ error: 'Task has no worktree' });
-      return;
-    }
-    const params = req.params as Record<string, string | string[]>;
-    const rawPath = params.path ?? params['0'] ?? '';
-    const relPath = Array.isArray(rawPath) ? rawPath.join('/') : rawPath;
-    try {
-      diffMod.safeResolvePath(cwd, relPath);
-    } catch {
-      res.status(400).json({ error: 'Invalid path' });
-      return;
-    }
-    try {
-      const { stdout } = await execFile('git', ['-C', cwd, 'rev-parse', 'HEAD']);
-      const headSha = stdout.trim();
-      // Capture the blob hash of the content actually reviewed (the working
-      // tree), so "changed since review" reacts to uncommitted edits too. Null
-      // when the file is gone (e.g. a reviewed deletion).
-      let blobSha: string | null = null;
-      try {
-        const { stdout: bs } = await execFile('git', ['-C', cwd, 'hash-object', '--', relPath]);
-        blobSha = bs.trim() || null;
-      } catch {
-        blobSha = null;
-      }
-      setReviewed(task.id, relPath, headSha, blobSha);
-      res.status(204).send();
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  app.delete('/api/tasks/:id/files/*path/reviewed', async (req: Request, res: Response) => {
-    const task = loadTaskOrFail(req, res);
-    if (!task) return;
-    const cwd = taskWorkingDir(task);
-    if (!cwd) {
-      res.status(400).json({ error: 'Task has no worktree' });
-      return;
-    }
-    const params = req.params as Record<string, string | string[]>;
-    const rawPath = params.path ?? params['0'] ?? '';
-    const relPath = Array.isArray(rawPath) ? rawPath.join('/') : rawPath;
-    try {
-      diffMod.safeResolvePath(cwd, relPath);
-    } catch {
-      res.status(400).json({ error: 'Invalid path' });
-      return;
-    }
-    clearReviewed(task.id, relPath);
-    res.status(204).send();
-  });
-
-  // ─── Inline review comments ────────────────────────────────────────────────
-
-  app.post('/api/tasks/:id/comments', async (req: Request, res: Response) => {
-    const task = loadTaskOrFail(req, res);
-    if (!task) return;
-    if (task.run_mode === 'scratch') {
-      res.status(400).json({ error: 'no repo for scratch task' });
-      return;
-    }
-    const cwd = taskWorkingDir(task);
-    if (!cwd) {
-      res.status(400).json({ error: 'Task has no worktree' });
-      return;
-    }
-    if (!fs.existsSync(cwd)) {
-      res.status(400).json({ error: 'Worktree no longer exists on disk' });
-      return;
-    }
-
-    const body = req.body as {
-      file_path?: unknown;
-      line?: unknown;
-      side?: unknown;
-      body?: unknown;
-      agent_id?: unknown;
-      anchor_commit_sha?: unknown;
-    };
-
-    const filePath = typeof body.file_path === 'string' ? body.file_path : '';
-    const lineRaw = body.line;
-    const side = body.side;
-    const commentBody = typeof body.body === 'string' ? body.body : '';
-    const agentId = typeof body.agent_id === 'string' && body.agent_id ? body.agent_id : null;
-    const anchorRaw = body.anchor_commit_sha;
-
-    if (!filePath) {
-      res.status(400).json({ error: 'file_path is required' });
-      return;
-    }
-    if (typeof lineRaw !== 'number' || !Number.isInteger(lineRaw) || lineRaw < 1) {
-      res.status(400).json({ error: 'line must be a positive integer' });
-      return;
-    }
-    if (side !== 'old' && side !== 'new') {
-      res.status(400).json({ error: "side must be 'old' or 'new'" });
-      return;
-    }
-    if (!commentBody.trim()) {
-      res.status(400).json({ error: 'body is required' });
-      return;
-    }
-    if (anchorRaw !== undefined && typeof anchorRaw !== 'string') {
-      res.status(400).json({ error: 'anchor_commit_sha must be a string' });
-      return;
-    }
-
-    try {
-      diffMod.safeResolvePath(cwd, filePath);
-    } catch {
-      res.status(400).json({ error: 'Invalid path' });
-      return;
-    }
-
-    if (!task.base_sha) {
-      res.status(400).json({ error: 'base_sha not available for this task' });
-      return;
-    }
-
-    try {
-      const row = await createInlineComment({
-        cwd,
-        task_id: task.id,
-        base_sha: task.base_sha,
-        file_path: filePath,
-        line: lineRaw as number,
-        side: side as 'old' | 'new',
-        body: commentBody,
-        agent_id: agentId,
-        anchor_commit_sha: anchorRaw as string | undefined,
-      });
-      res.status(201).json(row);
-    } catch (err) {
-      if (err instanceof ServiceError) {
-        res.status(err.status).json({ error: err.message });
-      } else {
-        res.status(500).json({ error: (err as Error).message });
-      }
-    }
-  });
-
-  app.get('/api/tasks/:id/comments', async (req: Request, res: Response) => {
-    const task = loadTaskOrFail(req, res);
-    if (!task) return;
-
-    const fileFilter = typeof req.query.file === 'string' ? req.query.file : undefined;
-    const rows = listComments(task.id, fileFilter ? { file: fileFilter } : undefined);
-
-    const cwd = taskWorkingDir(task);
-    const haveWorktree = !!cwd && fs.existsSync(cwd) && task.run_mode !== 'scratch';
-
-    if (!haveWorktree || !task.base_sha) {
-      res.json({
-        comments: rows.map((r) => ({ ...r, outdated: false })),
-        outdated_unavailable: true,
-      });
-      return;
-    }
-
-    try {
-      const map = await computeOutdated(cwd!, task.base_sha, rows);
-      res.json({
-        comments: rows.map((r) => ({ ...r, outdated: map.get(r.id) ?? false })),
-      });
-    } catch (err) {
-      apiLogger.warn(
-        { task_id: task.id, err: (err as Error).message },
-        'computeOutdated failed; returning comments without outdated flag',
-      );
-      res.json({
-        comments: rows.map((r) => ({ ...r, outdated: false })),
-        outdated_unavailable: true,
-      });
-    }
-  });
-
-  app.patch('/api/tasks/:id/comments/:cid', (req: Request, res: Response) => {
-    const task = loadTaskOrFail(req, res);
-    if (!task) return;
-
-    const cid = (req.params as Record<string, string>).cid;
-    const existing = getComment(cid);
-    if (!existing || existing.task_id !== task.id) {
-      res.status(404).json({ error: 'Comment not found' });
-      return;
-    }
-
-    // Refuse updates on already-published comments
-    if (existing.status === 'published') {
-      res.status(409).json({ error: 'Cannot update a published comment' });
-      return;
-    }
-
-    const body = req.body as {
-      resolved?: unknown;
-      body?: unknown;
-      status?: unknown;
-      bucket?: unknown;
-      kind?: unknown;
-      severity?: unknown;
-      existing_code?: unknown;
-      suggested_code?: unknown;
-      rejection_why?: unknown;
-    };
-    const hasResolved = body.resolved !== undefined;
-    const hasBody = body.body !== undefined;
-    const hasStatus = body.status !== undefined;
-    const hasExtended =
-      body.bucket !== undefined ||
-      body.kind !== undefined ||
-      body.severity !== undefined ||
-      body.existing_code !== undefined ||
-      body.suggested_code !== undefined;
-
-    const VALID_STATUSES = ['draft', 'accepted', 'rejected', 'stale'];
-
-    if (!hasResolved && !hasBody && !hasStatus && !hasExtended) {
-      res.status(400).json({ error: 'no fields to update' });
-      return;
-    }
-    if (hasResolved && typeof body.resolved !== 'boolean') {
-      res.status(400).json({ error: 'resolved must be a boolean' });
-      return;
-    }
-    if (hasBody && (typeof body.body !== 'string' || !body.body.trim())) {
-      res.status(400).json({ error: 'body must be a non-empty string' });
-      return;
-    }
-    if (hasStatus && !VALID_STATUSES.includes(body.status as string)) {
-      res.status(400).json({ error: `status must be one of ${VALID_STATUSES.join(', ')}` });
-      return;
-    }
-
-    let row = existing;
-
-    if (hasBody) {
-      const updated = updateCommentBody(cid, body.body as string);
-      if (updated) row = updated;
-    }
-    if (hasResolved) {
-      const updated = (body.resolved as boolean) ? resolveComment(cid) : unresolveComment(cid);
-      if (updated) row = updated;
-    }
-    if (hasStatus || hasExtended) {
-      const fields: import('./repositories/inline-comments.js').UpdateCommentFields = {};
-      if (hasStatus) fields.status = body.status as import('./types.js').CommentStatus;
-      if (body.bucket !== undefined)
-        fields.bucket = body.bucket as import('./types.js').CommentBucket | null;
-      if (body.kind !== undefined) fields.kind = body.kind as import('./types.js').CommentKind;
-      if (body.severity !== undefined)
-        fields.severity = body.severity as import('./types.js').CommentSeverity | null;
-      if (body.existing_code !== undefined)
-        fields.existing_code = body.existing_code as string | null;
-      if (body.suggested_code !== undefined)
-        fields.suggested_code = body.suggested_code as string | null;
-      const updated = updateCommentFields(cid, fields);
-      if (updated) row = updated;
-    }
-
-    // Capture rejection learning if status='rejected' and rejection_why provided
-    if (
-      body.status === 'rejected' &&
-      typeof body.rejection_why === 'string' &&
-      body.rejection_why.trim()
-    ) {
-      const repoPath = task.repo_path ?? '';
-      if (repoPath) {
-        addLearning({
-          repo_path: repoPath,
-          why: body.rejection_why.trim(),
-          created_from_comment_id: cid,
-        });
-      }
-    }
-
-    res.json(row);
-  });
-
-  app.delete('/api/tasks/:id/comments/:cid', (req: Request, res: Response) => {
-    const task = loadTaskOrFail(req, res);
-    if (!task) return;
-
-    const cid = (req.params as Record<string, string>).cid;
-    const existing = getComment(cid);
-    if (!existing || existing.task_id !== task.id) {
-      res.status(404).json({ error: 'Comment not found' });
-      return;
-    }
-
-    deleteComment(cid);
-    res.status(204).send();
-  });
-
   // ─── Worktrees (browser) ─────────────────────────────────────────────────
 
   app.get('/api/worktrees', (_req: Request, res: Response) => {
@@ -1665,285 +1083,6 @@ export function setupRoutes(app: Express): void {
     const limit = Math.min(parseInt((req.query.limit as string) || '50', 10), 200);
     const executions = getTaskHookExecutions(task.id, limit);
     res.json(executions);
-  });
-
-  // ─── Reviews inbox ───────────────────────────────────────────────────────────
-
-  // GET /api/reviews — list all auto_review tasks with aggregated counts
-  app.get('/api/reviews', (_req: Request, res: Response) => {
-    try {
-      res.json(listReviewsInbox());
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  // GET /api/reviews/:id — full detail for a single review task
-  app.get('/api/reviews/:id', (req: Request, res: Response) => {
-    try {
-      const detail = getReviewDetail((req.params as Record<string, string>).id);
-      if (!detail) {
-        res.status(404).json({ error: 'Review not found' });
-        return;
-      }
-      res.json(detail);
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  // POST /api/reviews — create an auto_review task for a GitHub PR URL.
-  // Idempotent: if a live (non-deleted, non-error) review task already exists
-  // for the same repo+PR, returns that instead of creating a duplicate.
-  app.post('/api/reviews', async (req: Request, res: Response) => {
-    const body = req.body as { pr_url?: unknown; repo_path?: unknown };
-    const prUrl = typeof body.pr_url === 'string' ? body.pr_url.trim() : '';
-    const bodyRepoPath = typeof body.repo_path === 'string' ? body.repo_path.trim() : '';
-
-    // Parse GitHub PR URL
-    const prMatch = prUrl.match(/github\.com\/([^/]+)\/([^/]+)\/pull\/(\d+)/);
-    if (!prMatch) {
-      res.status(400).json({ error: 'invalid pr_url' });
-      return;
-    }
-    const [, owner, repo, numberStr] = prMatch;
-    const number = parseInt(numberStr, 10);
-    const ownerRepo = `${owner}/${repo}`;
-
-    // Resolve the local repo path
-    let repoPath = bodyRepoPath;
-    if (!repoPath) {
-      const rows = listTrackedRepoPaths();
-
-      for (const row of rows) {
-        const candidatePath = row.repo_path;
-        try {
-          const { stdout } = await execFile('git', [
-            '-C',
-            candidatePath,
-            'remote',
-            'get-url',
-            'origin',
-          ]);
-          const remoteUrl = stdout.trim();
-          // Handle both ssh (git@github.com:owner/repo.git) and https
-          const sshMatch = remoteUrl.match(/git@github\.com:([^/]+\/[^/]+?)(?:\.git)?$/);
-          const httpsMatch = remoteUrl.match(/github\.com\/([^/]+\/[^/]+?)(?:\.git)?$/);
-          const remoteOwnerRepo = (sshMatch?.[1] ?? httpsMatch?.[1] ?? '').toLowerCase();
-          if (remoteOwnerRepo === ownerRepo.toLowerCase()) {
-            repoPath = candidatePath;
-            break;
-          }
-        } catch {
-          // skip this candidate
-        }
-      }
-    }
-
-    if (!repoPath) {
-      res.status(400).json({
-        error: `could not resolve a local repo for ${ownerRepo}; pass repo_path`,
-      });
-      return;
-    }
-
-    // Dedup: check for an existing live review task for this repo+PR
-    const existing = findExistingReviewTask(repoPath, number);
-
-    if (existing) {
-      res.status(200).json({ id: existing.id, reused: true });
-      return;
-    }
-
-    // Fetch PR metadata via gh CLI
-    let pr: {
-      title: string;
-      headRefOid: string;
-      baseRefName: string;
-      author: { login: string } | null;
-      state: string;
-      url: string;
-    };
-    try {
-      const { stdout } = await execFile(
-        'gh',
-        [
-          'pr',
-          'view',
-          String(number),
-          '--repo',
-          ownerRepo,
-          '--json',
-          'title,headRefOid,baseRefName,author,state,url',
-        ],
-        { cwd: repoPath },
-      );
-      pr = JSON.parse(stdout) as typeof pr;
-    } catch (err) {
-      res.status(400).json({ error: `failed to fetch PR metadata: ${(err as Error).message}` });
-      return;
-    }
-
-    if (pr.state !== 'OPEN') {
-      res.status(400).json({ error: `PR #${number} is ${pr.state}` });
-      return;
-    }
-
-    // Create the review task (service owns the build+insert+broadcast+startTask tail).
-    const { id } = await createReviewTaskFromPr({
-      repo_path: repoPath,
-      pr_number: number,
-      pr_url: pr.url,
-      pr_head_sha: pr.headRefOid,
-      base_branch: pr.baseRefName,
-      title: pr.title,
-      author: pr.author?.login ?? null,
-      requested_at: new Date().toISOString(),
-    });
-
-    apiLogger.info(
-      { task_id: id, pr_number: number, repo: ownerRepo, repo_path: repoPath },
-      'review create task created',
-    );
-
-    res.status(201).json({ id, reused: false });
-  });
-
-  // ─── Review runs ─────────────────────────────────────────────────────────────
-
-  // PATCH /api/tasks/:id/review-runs/:rid/walkthrough — deep-merge walkthrough
-  app.patch('/api/tasks/:id/review-runs/:rid/walkthrough', (req: Request, res: Response) => {
-    const task = loadTaskOrFail(req, res);
-    if (!task) return;
-
-    const params = req.params as Record<string, string>;
-    const rid = params.rid;
-
-    const run = getReviewRun(rid);
-    if (!run || run.task_id !== task.id) {
-      res.status(404).json({ error: 'Review run not found' });
-      return;
-    }
-
-    // Refuse if a published_reviews row exists for this run's pr_head_sha
-    const published = listPublishedReviews(task.id);
-    const alreadyPublished = published.some((p) => p.head_sha === run.pr_head_sha);
-    if (alreadyPublished) {
-      res.status(409).json({ error: 'Review already published for this head SHA' });
-      return;
-    }
-
-    // Deep-merge incoming body into existing walkthrough JSON
-    const existing = run.walkthrough ? JSON.parse(run.walkthrough) : {};
-    const incoming = req.body as Record<string, unknown>;
-    const merged = deepMerge(existing, incoming);
-    setWalkthrough(rid, JSON.stringify(merged));
-
-    const updated = getReviewRun(rid);
-    res.json(updated);
-  });
-
-  // POST /api/tasks/:id/review-runs — trigger a manual re-review
-  app.post('/api/tasks/:id/review-runs', async (req: Request, res: Response) => {
-    const task = loadTaskOrFail(req, res);
-    if (!task) return;
-
-    // 409 if a run with status='running' already exists for this task
-    const currentRun = getCurrentRun(task.id);
-    if (currentRun?.status === 'running') {
-      res.status(409).json({ error: 'A review run is already in progress' });
-      return;
-    }
-
-    try {
-      await triggerReviewRun(task);
-      res.status(202).json({ ok: true });
-    } catch (err) {
-      res.status(500).json({ error: (err as Error).message });
-    }
-  });
-
-  // POST /api/tasks/:id/publish-review — publish accepted draft comments to GitHub
-  app.post('/api/tasks/:id/publish-review', async (req: Request, res: Response) => {
-    const task = loadTaskOrFail(req, res);
-    if (!task) return;
-
-    const body = req.body as { verdict?: unknown; review_body?: unknown };
-    const verdict = body.verdict ?? 'COMMENT';
-
-    if (!['COMMENT', 'APPROVE', 'REQUEST_CHANGES'].includes(verdict as string)) {
-      res.status(400).json({ error: 'verdict must be one of COMMENT, APPROVE, REQUEST_CHANGES' });
-      return;
-    }
-
-    try {
-      const { publishReview } = await import('./publish-review.js');
-      const result = await publishReview(
-        task.id,
-        verdict as import('./types.js').PublishedReviewVerdict,
-        typeof body.review_body === 'string' ? body.review_body : '',
-      );
-      res.json(result);
-    } catch (err) {
-      const msg = (err as Error).message;
-      if (msg.includes('No accepted comments')) {
-        res.status(400).json({ error: msg });
-      } else {
-        res.status(500).json({ error: msg });
-      }
-    }
-  });
-
-  // POST /api/tasks/:taskId/review — manually trigger a review for this task.
-  // Creates an auto_review task pointing back at the source via review_of_task_id,
-  // or returns the existing review when one is already present.
-  app.post('/api/tasks/:taskId/review', async (req: Request, res: Response) => {
-    const taskId = (req.params as Record<string, string>).taskId;
-    const task = getTaskRepo(taskId);
-    if (!task) {
-      res.status(404).json({ error: 'Task not found' });
-      return;
-    }
-
-    if (!task.branch || !task.worktree) {
-      res.status(400).json({ error: 'Start the task first' });
-      return;
-    }
-
-    const existingId = lookupExistingReviewId(task);
-    if (existingId) {
-      res.status(200).json({ id: existingId, action: 'existing' });
-      return;
-    }
-
-    // PR-less source: capture the current HEAD of the source worktree so the
-    // review agent has a concrete sha to diff base..head against.
-    let prHeadSha = task.pr_head_sha;
-    if (!prHeadSha) {
-      const cwd = taskWorkingDir(task);
-      try {
-        const { stdout } = await execFile('git', ['-C', cwd!, 'rev-parse', 'HEAD']);
-        prHeadSha = stdout.trim();
-      } catch (err) {
-        res.status(500).json({ error: `failed to resolve HEAD: ${(err as Error).message}` });
-        return;
-      }
-    }
-
-    const { id: newId } = await createManualReview({
-      source_task_id: task.id,
-      source_title: task.title,
-      repo_path: task.repo_path,
-      branch: task.branch,
-      base_branch: task.base_branch,
-      base_sha: task.base_sha ?? null,
-      pr_head_sha: prHeadSha,
-      pr_url: task.pr_url ?? null,
-      pr_number: task.pr_number ?? null,
-      requested_at: new Date().toISOString(),
-    });
-
-    res.status(201).json({ id: newId, action: 'created' });
   });
 
   // ─── Test-only seed endpoint ─────────────────────────────────────────────────
